Got something to show finally in lovelace!
This was not easy. I had to try a million ways to import
it and then do some hackery of my plotly library as
described in https://github.com/plotly/angular-plotly.js/issues/76#issuecomment-526872398
but finally I have something. It's not perfect and doesn't let
me mouseover the data or anything, but it's certainly something!!